rename internal to HistData
